+++ streamx-monitor/streamx-monitor-service/src/main/java/com/streamxhub/monitor/core/entity/Application.java
@@ -93,7 +93,7 @@ public class Application implements Serializable {
     private transient String userName;
 
     private transient String config;
-    private transient Integer configVersion;
+    private transient Long configVersion;
     private transient String confPath;
     private transient Integer format;
     private transient String savePoint;
